Avoid name collisions with concurrent tests
* Helpers + test integration for creating unique, repeatable values to inject in template strings
