+++ resource_obmcs_core_vnic_attachment_test.go
@@ -8,6 +8,7 @@ import (
 	"regexp"
 
 	"fmt"
+
 	"github.com/hashicorp/terraform/helper/resource"
 	"github.com/hashicorp/terraform/terraform"
 	"github.com/oracle/bmcs-go-sdk"
